api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.30
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.7.1
pydantic-settings==2.2.1